crawler/Youtube-Download/requirements.txt
yt-dlp
pyyaml
ffmpeg
pydub
webvtt-py